==> 01-collect-data.ipynb <==
# Collect Data
A primeira etapa para o desenvolvimento do projeto é a coleta de dados. Os dados estão armazenados em tabelas distintas no BigQuery, e para obte-los será necessário realizar ela série de operações em SQL para manipular os dados e extrair as métricas dos e-mails enviados.

## Imports
Bibliotecas e ferramentas necessárias para o desenvolvimento dessa etapa.

import pandas as pd
from google.oauth2.credentials import Credentials
from google.cloud import bigquery
import os

GCP_CREDENTIALS       = os.getenv('GCP_CREDENTIALS', "keys.json")
PROJECT_ID            = os.getenv('PROJECT_ID', "etus-media-data-staging")

## Authentication
Autenticação para acessar os dados do BigQuery.

creds = Credentials.from_authorized_user_file(GCP_CREDENTIALS)
client = bigquery.Client(PROJECT_ID, credentials=creds)

## Extract Data
Coletar dados no BigQuery e salvar localmente.

with open('./query/query.sql', 'r') as file:
    query = file.read()

df = client.query(query).result().to_dataframe()
df.to_csv('./data/raw_data.csv', index=False)
