==> 04-split-data.ipynb <==
# Split Data into Train and Test
Separar os dados, em um conjunto de treino e outro para teste.

## Imports

import pandas as pd
from sklearn.model_selection import train_test_split
import json

## Read Data and Split

def split_and_save_dataset(
        file_path,
        test_size=0.2,
        val_size=0.1,
        random_state=None, 
        train_file_path='./data/train.csv', 
        test_file_path='./data/test.csv',
        val_file_path='./data/val.csv'
    ):
    
    df = pd.read_csv(file_path)
    
    df['EMBEDDING_TFIDF'] = df['EMBEDDING_TFIDF'].astype(str)
    df['EMBEDDING_WORD2VEC'] = df['EMBEDDING_WORD2VEC'].astype(str)
    df['EMBEDDING_OPENAI'] = df['EMBEDDING_OPENAI'].astype(str)
    
    train_df, test_df = train_test_split(df, test_size=test_size, random_state=random_state)
    train_df, val_df = train_test_split(train_df, test_size=val_size, random_state=random_state)
    
    train_df.to_csv(train_file_path, index=False)
    test_df.to_csv(test_file_path, index=False)
    val_df.to_csv(val_file_path, index=False)
    
    print(f"Conjunto de treino salvo em '{train_file_path}'")
    print(f"Conjunto de teste salvo em '{test_file_path}'")
    print(f"Conjunto de validação salvo em '{val_file_path}'")

file_path = './data/data_preprocessed_embeddings.csv'

split_and_save_dataset(file_path, test_size=0.20, val_size=0.10, random_state=42)

## Train Data

train = pd.read_csv('./data/train.csv')
train

## Test Data

test = pd.read_csv('./data/test.csv')
test
